+++ Sources/Atlas.h
@@ -22,7 +22,7 @@ namespace aga
         Atlas (const std::string& path);
         ~Atlas ();
 
-        void DrawRegion (const std::string& name, int x, int y, float scaleX, float scaleY, float rotation);
+        void DrawRegion (const std::string& name, float x, float y, float scaleX, float scaleY, float rotation);
         ALLEGRO_BITMAP* GetImage ();
         AtlasRegion& GetRegion (const std::string& name);
         std::vector<AtlasRegion>& GetRegions ();
